unix home directory - give precedence to $HOME
Give precedence to the $HOME environment variable of the password
database.  The credential helper may be running in a user namespace
where the database is likely tricked into looking up a wrong user
since the current UID is different.

Giving precedence to $HOME will fix an issue in Podman [1] where the
credential helper is erroring out with EPERMs since it uses the root's
home directory.

[1] https://lists.podman.io/archives/list/podman@lists.podman.io/thread/CZY7J34AZHXZOK34Y4JNNIOVR62N43XO/

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>