Automatically initiate browser-based reauth when needed
This change modifies the auth refresh flow to detect the specific error
response that indicates a reauth is required ("invalid_rapt"), and
initiates a new browser-based auth flow in that case.

This specific error indicates that a human user's auth is used and that
the user's GCP organization requires periodic browser-based auth flows,
for example to impose 2FA requirements.

I don't anticipate this error response will be seen using service
account credentials, or when a human user's credentials are involved
without access to a browser. In either case, the "invalid_rapt" response
indicates the user must open a browser to proceed, e.g., using
`gcr-login` or `gcloud auth login`, so this just automates the
initiation of that flow.

Signed-off-by: Jason Hall <jason@chainguard.dev>